Partially revert 8036ad5 to fix #242
This partially reverts commit 8036ad5 which broke the audio in dynarecs on CT Special Forces 3.  Essentially restoring the setting of the IWRAM/EWRAM thresholds back to the original positions in the code, but keeping the other changes in 8036ad5.

Applying this PR restores the working audio in this game.

Up to @davidgfnet if this is satisfactory to commit or not without further investigation.