DRAFT - OAM Order Hijack Support (to fix #227 and similar)
This is my attempt to support OAM Order Hijacking.  The cause and effect of this is outlined in issue #227.

The issue that I see for gpsp in addressing this issue is that we store the Layer Priority at the Object/BG levels only.  In order to correctly deal with OAM Order Hijacking in the same way as the hardware, we need to be able to check priority at a pixel-level and act accordingly.

My solution here is - 

a) to check for OAM Order Hijacking at a row-level when we are ordering our objects (in order_obj), and mark each applicable row in the same way we do for rows that contain transparent objects.
b) when we come to render a scanline marked as hijacked, generate an OBJ-only scanline buffer first by rendering all available OBJ layers to this buffer in u16/INDXCOLOR mode, which gives us the most flexibility later on
c) when rendering 8-bit per pixel OBJ tiles, we check 4 pixels at a time for transparency (tilepix).  In the current code, if all 4 pixels are transparent (i.e. 0), we move on to the next 4 pixels to complete the tile.  But in my PR, we instead check each pixel to see if we have a non-zero value in the OBJ scanline buffer we created.  If we do, we render the value to dest_ptr using the selected render mode.

I have tried to build in some optimisations here such as only creating the OBJ scanline buffer on scanlines affected by a hijack and only comparing to the scanline buffer when we have 4 consecutive transparent pixels as described above.  The latter is sufficient to address the issues identified in Golden Sun as far as I can see, but complete/correct emulation would be to compare against EVERY transparent pixel.  

Further ToDos:
- 4bpp support
- Affine/Mosaic Objects support

So this PR is still in draft at the moment for more discussion/testing before committing.