Separate MAX_BLOCK_SIZE / MAX_EXITS for arm/thumb mode and reduce arm values
This improves performance on Camelot games (specifically Mario Tennis / Mario Golf) and solves issue #151.  Theory is that by reducing the block sizes/exits for arm mode that this reduces the number of times the dynarec caches need to be flushed per frame.  Impact of this change needs to be assessed for all games though.