Increase frequency of Interrupts/DMA instance checks during scanline
This resolves issue #224.  All games in this issue (all from the same developer) generate a high number of DMA/IRQ requests per scanline relative to other games.  Theory is that the default frequency (about once or twice per scanline) isn't enough for these to be fired in a timely fashion for these games, resulting in the graphics corruption reported in the issue.

This PR significantly reduces the number of cycles that are run before update_gba is called to check for DMA/IRQ status, which resolves the corrupted graphics.  Casual testing with other games hasn't shown any side affects but full impact testing with other games will be needed.